refactor: rename embodied rollout result and reorg vlm dataset
