+++ docs/source-en/rst_source/guides/data_collection.rst
@@ -300,7 +300,7 @@ Core Components
 1. Initialise ``RealWorldEnv`` and ``TrajectoryReplayBuffer``.
 2. Loop over steps, reading the SpaceMouse intervention action from
    ``info["intervene_action"]``.
-3. Construct a ``ChunkStepResult`` and append it to ``EmbodiedRolloutResult``.
+3. Construct a ``ChunkStepResult`` and append it to ``EmbodiedTrajectoryBuilder``.
 4. When an episode ends (``done=True``) with reward ``>= 0.5``, count it as a
    success and write the trajectory to the buffer.
 5. Stop automatically once ``num_data_episodes`` successes have been collected
