+++ examples/agent/coding_online_rl/main_coding_rl_llm_judge.py
@@ -19,7 +19,7 @@ import torch.multiprocessing as mp
 from omegaconf.omegaconf import OmegaConf
 
 from rlinf.config import validate_cfg
-from rlinf.data.datasets.reasoning import create_rl_dataset
+from rlinf.data.datasets.reasoning import create_reasoning_datasets
 from rlinf.models.tokenization.hf import hf_tokenizer
 from rlinf.runners.reasoning_runner import ReasoningRunner
 from rlinf.scheduler import Cluster
@@ -84,7 +84,7 @@ def main(cfg) -> None:
     )
 
     tokenizer = hf_tokenizer(cfg.actor.tokenizer.tokenizer_model)
-    train_ds, val_ds = create_rl_dataset(cfg, tokenizer)
+    train_ds, val_ds = create_reasoning_datasets(cfg, tokenizer)
 
     runner = ReasoningRunner(
         cfg=cfg,
