+++ examples/agent/searchr1/eval.py
@@ -22,7 +22,7 @@ from rlinf.agents.searchr1.eval_runner import Searchr1AgentEvalRunner as AgentEv
 from rlinf.agents.searchr1.search_tool_worker import SearchToolWorker
 from rlinf.agents.searchr1.searchr1_agent_loop import Searchr1AgentLoopWorker
 from rlinf.config import validate_cfg
-from rlinf.data.datasets.reasoning import create_rl_dataset
+from rlinf.data.datasets.reasoning import create_reasoning_datasets
 from rlinf.models.tokenization.hf import hf_tokenizer
 from rlinf.scheduler import Cluster, NodePlacementStrategy
 from rlinf.utils.placement import ModelParallelEvalComponentPlacement
@@ -75,7 +75,7 @@ def main(cfg) -> None:
 
     # Dataset
     tokenizer = hf_tokenizer(cfg.rollout.model.model_path)
-    train_ds, val_ds = create_rl_dataset(cfg, tokenizer)
+    train_ds, val_ds = create_reasoning_datasets(cfg, tokenizer)
 
     # Tool workers group
     singleton_tool_placement = NodePlacementStrategy([0])
