+++ rlinf/data/storage/replay/dataset.py
@@ -28,7 +28,23 @@ logger = get_logger()
 
 
 class ReplayBufferDataset(IterableDataset):
-    """Dataset that samples batches from replay and demonstration buffers."""
+    """Dataset that samples batches from replay and demonstration buffers.
+
+    This dataset provides an infinite iterator that yields batches sampled from
+    a replay buffer and optionally a demonstration buffer. When both buffers are
+    provided, batches are composed of half replay samples and half demonstration
+    samples.
+
+    Attributes:
+        replay_buffer: Buffer storing online rollout trajectories.
+        demo_buffer: Optional buffer storing offline demonstration trajectories
+            and online human-in-the-loop trajectories.
+        min_replay_buffer_size: Minimum number of samples required in replay
+            buffer before sampling begins.
+        min_demo_buffer_size: Minimum number of samples required in demo buffer
+            before sampling begins (if demo_buffer is provided).
+        batch_size: Total number of samples per batch.
+    """
 
     def __init__(
         self,
@@ -39,6 +55,20 @@ class ReplayBufferDataset(IterableDataset):
         min_demo_buffer_size: int,
         **kwargs: Any,
     ) -> None:
+        """Initializes the ReplayBufferDataset.
+
+        Args:
+            replay_buffer: Buffer storing online rollout trajectories.
+            demo_buffer: Optional buffer storing demonstration trajectories.
+                If None, only replay buffer is used.
+            batch_size: Total number of samples per batch. When demo_buffer is
+                provided, batch_size // 2 samples come from each buffer.
+            min_replay_buffer_size: Minimum number of samples required in replay
+                buffer before sampling begins.
+            min_demo_buffer_size: Minimum number of samples required in demo
+                buffer before sampling begins (ignored if demo_buffer is None).
+            **kwargs: Additional keyword arguments (unused, for compatibility).
+        """
         self.replay_buffer = replay_buffer
         self.demo_buffer = demo_buffer
         self.min_replay_buffer_size = min_replay_buffer_size
@@ -46,6 +76,16 @@ class ReplayBufferDataset(IterableDataset):
         self.batch_size = batch_size
 
     def __iter__(self) -> Iterator[dict[str, torch.Tensor]]:
+        """Returns an infinite iterator that yields batches.
+
+        Waits until both buffers (if demo_buffer is provided) reach their
+        minimum size requirements before yielding batches. When ready, samples
+        from replay buffer only or from both replay and demo buffers.
+
+        Yields:
+            Batch dictionary containing sampled trajectories. Keys and structure
+            depend on the buffer's trajectory format.
+        """
         while True:
             is_ready = True
             if not self.replay_buffer.is_ready(self.min_replay_buffer_size):
@@ -65,15 +105,34 @@ class ReplayBufferDataset(IterableDataset):
                 yield batch
 
     def close(self) -> None:
+        """Releases references to replay and demo buffers."""
         del self.replay_buffer
         del self.demo_buffer
 
     def __del__(self) -> None:
+        """Destructor that ensures buffers are cleaned up."""
         self.close()
 
 
 class PreloadReplayBufferDataset(ReplayBufferDataset):
-    """Dataset that prefetches batches in a background thread."""
+    """Dataset that prefetches batches from replay and demo buffers in background.
+
+    This dataset extends ReplayBufferDataset by prefetching batches in a
+    background thread, which can improve throughput by overlapping sampling
+    with training. Batches are stored in a queue of configurable size.
+
+    Attributes:
+        replay_buffer: Buffer storing online rollout trajectories.
+        demo_buffer: Optional buffer storing demonstration trajectories.
+        min_replay_buffer_size: Minimum number of samples required in replay
+            buffer before sampling begins.
+        min_demo_buffer_size: Minimum number of samples required in demo buffer
+            before sampling begins (if demo_buffer is provided).
+        batch_size: Total number of samples per batch.
+        prefetch_size: Maximum number of batches to prefetch and store in queue.
+        preload_queue: Queue holding prefetched batches.
+        sample_thread: Background thread that samples batches.
+    """
 
     def __init__(
         self,
@@ -84,6 +143,21 @@ class PreloadReplayBufferDataset(ReplayBufferDataset):
         min_demo_buffer_size: int,
         prefetch_size: int = 5,
     ) -> None:
+        """Initializes the PreloadReplayBufferDataset.
+
+        Args:
+            replay_buffer: Buffer storing online rollout trajectories.
+            demo_buffer: Optional buffer storing demonstration trajectories.
+                If None, only replay buffer is used.
+            batch_size: Total number of samples per batch. When demo_buffer is
+                provided, batch_size // 2 samples come from each buffer.
+            min_replay_buffer_size: Minimum number of samples required in replay
+                buffer before sampling begins.
+            min_demo_buffer_size: Minimum number of samples required in demo
+                buffer before sampling begins (ignored if demo_buffer is None).
+            prefetch_size: Maximum number of batches to prefetch and store in
+                the queue. Defaults to 10.
+        """
         self._stop_event = threading.Event()
 
         self.replay_buffer = replay_buffer
@@ -100,6 +174,13 @@ class PreloadReplayBufferDataset(ReplayBufferDataset):
         self._exception = None
 
     def _sample_buffer(self) -> None:
+        """Background thread target that continuously samples batches.
+
+        Runs in a loop until stop event is set. Waits for buffers to be ready,
+        samples batches, and puts them in the preload queue. If the queue is
+        full, skips the sample and retries. Sleeps when buffers are not ready
+        or when errors occur.
+        """
         while not self._stop_event.is_set():
             if self.preload_queue.full():
                 time.sleep(0.1)
@@ -137,6 +218,15 @@ class PreloadReplayBufferDataset(ReplayBufferDataset):
                 break
 
     def __iter__(self) -> Iterator[dict[str, torch.Tensor]]:
+        """Returns an iterator that yields prefetched batches.
+
+        Starts the background sampling thread on first call. Retrieves batches
+        from the preload queue and yields them. Stops when the stop event is set.
+
+        Yields:
+            Batch dictionary containing sampled trajectories. Keys and structure
+            depend on the buffer's trajectory format.
+        """
         if self.sample_thread is None:
             self.sample_thread = threading.Thread(
                 target=self._sample_buffer, daemon=True
@@ -149,6 +239,7 @@ class PreloadReplayBufferDataset(ReplayBufferDataset):
                 yield batch
             except queue.Empty:
                 if self._stop_event.is_set():
+                    # Check if thread died with exception
                     if hasattr(self, "_exception"):
                         raise RuntimeError(
                             "Sampling thread failed"
@@ -157,6 +248,11 @@ class PreloadReplayBufferDataset(ReplayBufferDataset):
                 continue
 
     def close(self) -> None:
+        """Stops the background sampling thread and cleans up resources.
+
+        Sets the stop event and waits up to 10 seconds for the sampling thread
+        to terminate. Logs a warning if the thread does not terminate in time.
+        """
         self._stop_event.set()
 
         thread_timeout = 10
@@ -168,6 +264,7 @@ class PreloadReplayBufferDataset(ReplayBufferDataset):
                 )
 
     def __del__(self) -> None:
+        """Destructor that ensures the sampling thread is stopped."""
         if not self._stop_event.is_set():
             self.close()
 
@@ -175,4 +272,15 @@ class PreloadReplayBufferDataset(ReplayBufferDataset):
 def replay_buffer_collate_fn(
     batch: list[dict[str, torch.Tensor]],
 ) -> dict[str, torch.Tensor]:
+    """Collate function for DataLoader that returns the first batch element.
+
+    Since the dataset already yields complete batches, this function simply
+    extracts the batch from the list wrapper added by DataLoader.
+
+    Args:
+        batch: List containing a single batch dictionary.
+
+    Returns:
+        The unwrapped batch dictionary.
+    """
     return batch[0]
