+++ rlinf/workers/sft/megatron_vlm_sft_worker.py
@@ -300,20 +300,18 @@ class MegatronVlmSftWorker(MegatronSftWorker):
         ]:
             from torch.utils.data import DataLoader, DistributedSampler
 
-            from rlinf.data.datasets.reasoning import sft_collate_fn
-            from rlinf.data.datasets.vlm import VLMDatasetRegistry
+            from rlinf.data.datasets.vlm import collate_fn, create_vlm_datasets
 
             if not hasattr(self, "tokenizer"):
                 self.tokenizer = self.build_tokenizer()
 
-            dataset_name = self.cfg.data.get("dataset_name", "robo2vlmsft")
-            train_dataset = VLMDatasetRegistry.create(
-                dataset_name,
+            train_dataset, _ = create_vlm_datasets(
+                self.cfg,
+                self.tokenizer,
                 data_paths=data_paths,
-                config=self.cfg,
-                tokenizer=self.tokenizer,
                 eval_dataset=eval_dataset,
             )
+            dataset_name = self.cfg.data.get("dataset_name", "robo2vlmsft")
 
             import torch.distributed as dist
 
@@ -343,7 +341,7 @@ class MegatronVlmSftWorker(MegatronSftWorker):
                 shuffle=(sampler is None),
                 num_workers=self.cfg.data.get("num_workers", 4),
                 drop_last=True,
-                collate_fn=sft_collate_fn,
+                collate_fn=collate_fn,
             )
             logging.info(
                 f"Build data loader from {data_paths} with {len(train_dataset)} samples"
